requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
appdirs==1.4.4            # via black
atomicwrites==1.4.0       # via pytest
attrs==21.2.0             # via pytest
black==21.5b1             # via -r requirements.in
certifi==2020.12.5        # via py2neo, requests
cffi==1.14.5              # via cryptography
chardet==4.0.0            # via requests
click==7.1.2              # via black, uvicorn
colorama==0.4.4           # via pytest
cryptography==3.4.7       # via py2neo
docker==5.0.0             # via py2neo
english==2020.7.0         # via py2neo
fastapi==0.63.0           # via -r requirements.in
flake8==3.9.2             # via -r requirements.in
h11==0.12.0               # via uvicorn
idna==2.10                # via requests
importlib-metadata==4.0.1  # via flake8, pluggy, pytest
iniconfig==1.1.1          # via pytest
mccabe==0.6.1             # via flake8
mock==4.0.3               # via -r requirements.in
monotonic==1.6            # via py2neo
mypy-extensions==0.4.3    # via black
neotime==1.7.4            # via py2neo
packaging==20.9           # via py2neo, pytest
pansi==2020.7.3           # via py2neo
pathspec==0.8.1           # via black
pluggy==0.13.1            # via pytest
prompt-toolkit==2.0.10    # via py2neo
py2neo==2021.0.1          # via -r requirements.in
py==1.10.0                # via pytest
pycodestyle==2.7.0        # via flake8
pycparser==2.20           # via cffi
pydantic==1.8.1           # via fastapi
pyflakes==2.3.1           # via flake8
pygments==2.8.1           # via py2neo
pyparsing==2.4.7          # via packaging
pytest==6.2.4             # via -r requirements.in
pytz==2021.1              # via neotime, py2neo
pyyaml==5.4.1             # via -r requirements.in
regex==2021.4.4           # via black
requests==2.25.1          # via docker
six==1.15.0               # via english, neotime, pansi, prompt-toolkit, py2neo, websocket-client
starlette==0.13.6         # via fastapi
toml==0.10.2              # via black, pytest
typed-ast==1.4.3          # via black
typing-extensions==3.7.4.3  # via black, importlib-metadata, pydantic, uvicorn
urllib3==1.26.4           # via py2neo, requests
uvicorn==0.13.4           # via -r requirements.in
wcwidth==0.2.5            # via prompt-toolkit
websocket-client==0.58.0  # via docker
zipp==3.4.1               # via importlib-metadata
